Changes the grid div dimensions/number
Function changed to recalculate the dimensions of the grid div and to change the number of the grid to reflect the user input being the grid dimension; i.e. userInput = 16 => 16x16 grid.

Dimensions are changed to reflect that any given side of a grid div is the length of the container divided by the userinput number.